fix!: Use double instead of float in map API
Java numbers like 1.2 are by default doubles and having floats in the API only makes it harder to use

+++ vaadin-map-flow-parent/vaadin-map-flow/src/main/java/com/vaadin/flow/component/map/MapBase.java
@@ -132,8 +132,8 @@ public abstract class MapBase extends Component implements HasSize, HasTheme {
         // move end event to update view state to the latest values received
         // from the client
         addViewMoveEndEventListener(event -> {
-            float rotation = event.getRotation();
-            float zoom = event.getZoom();
+            double rotation = event.getRotation();
+            double zoom = event.getZoom();
             Coordinate center = event.getCenter();
             Extent extent = event.getExtent();
             getView().updateInternalViewState(center, rotation, zoom, extent);
